test: isolate fused scan performance harness
